--- ROSBLOCKS/src/app/blockly/blockly.component.css
@@ -1,10 +1,0 @@
-#searchBox {
-    left: 22px;
-    position: absolute;
-    margin-bottom: 10px;
-    width: 190px !important;
-    padding: 5px;
-    top: 25px;
-    z-index: 999;
-  }
-
