--- ROSBLOCKS/src/app/blockly/blockly.component.html
@@ -1,10 +1,0 @@
-<div class="blockly-container">
-    <input
-        id="searchBox"
-        type="text"
-        placeholder="Search for a block..."
-        (input)="onSearch($event)"
-        style="margin-bottom: 10px; width: 100%; padding: 5px;"
-    />
-    <div id="blocklyDiv" style="height: 500px; width: 100%;"></div>
-</div>
